Assignment of positions is now based on player id.
It was previously depending on the index of a player's position
input corresponding with that player's index in the storage array.
This would have broken when we introduce the ability to mark
players as inactive and should make it easier to reuse the same
logic when editing of Games is implemented.